services/face-recognition-service/requirements.txt

facenet-pytorch
pika
python-dotenv
opencv-python-headless
pymilvus